requirements.txt
django==2.2.4
django-annoying==0.10.4
django-appconf==1.0.3
django-imagekit==4.0.2
gunicorn==19.9.0
pilkit==2.0
pillow==6.1.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
